#250 [refactor] Model Controller 조회/등록 컨트롤러 2가지로 분리
